# src/server/swagger.yaml
openapi: 3.0.0
info:
  title: Kroombox API Gateway
  description: Dokumentasi Lengkap API Gateway untuk Kroombox, mencakup manajemen klien, proxy, metrik GPU, event realtime, dan panel admin.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local server
  - url: "{appUrl}"
    description: Production server
    variables:
      appUrl:
        default: https://api.kroombox.com
        description: Base URL untuk Production
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: 'Gunakan format: Bearer {API_KEY}'
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /api/auth/token:
    post:
      tags: [Authentication]
      summary: Generate Token Akses Klien berdasarkan API Key
      responses:
        '200':
          description: Sukses
        '401':
          description: Unauthorized
...